Separating integration from unit tests
